Add files via upload
Using data processed in the Coller Lab using 15 studies of quiescence that assigns a number to gene representing its up-regulation or down-regulation with quiescence in different cancer samples. I compared these genes to data from the MSK MetTropism study that contains the frequencies of mutation, structural variation, and CNAs in cancer cells.